confirm event changes , noti connexted on user home screen see all
